move auth to msg module, export msg module as package entry point
